% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_test.R
\name{auc_test}
\alias{auc_test}
\title{Compare two AUC values from ROC or PRC curves}
\usage{
auc_test(
  data,
  y,
  x1,
  x2,
  nboot = 1000,
  seed = NULL,
  prc = FALSE,
  stratified = TRUE,
  verbose = TRUE,
  alternative = c("two.sided", "less", "greater")
)
}
\arguments{
\item{data}{dataset to fit ROC or PRC curves to}

\item{y}{Character string name of outcome}

\item{x1}{Character string or vector containing predictor name(s) for first
model}

\item{x2}{Character string or vector containing predictor name(s) for second
model}

\item{nboot}{Number of bootstrap iterations}

\item{seed}{Numeric value for seed}

\item{prc}{Logical, determines whether to fit PR or ROC curves}

\item{stratified}{Logical. Determines whether to retain class proportions
during bootstrapped resampling}

\item{verbose}{Logical. Determines whether to show progress bar}

\item{alternative}{Character, determines what kind of alternative hypothesis}
}
\value{
D-statistic and associated p-value
}
\description{
Compare two AUC values from ROC or PRC curves
}
\examples{
1 + 1
}
