% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROCcurve.R
\name{plot.ROCcurve}
\alias{plot.ROCcurve}
\title{Plotting function for ROC and PRC curves}
\usage{
\method{plot}{ROCcurve}(x, ...)
}
\arguments{
\item{x}{ROCcurve object}

\item{...}{Additional arguments}
}
\value{
An ROC or PRC plot
}
\description{
Plotting function for ROC and PRC curves
}
\examples{
roc <- ROCcurve(mtcars$vs, mtcars$mpg)
prc <- ROCcurve(mtcars$vs, mtcars$mpg, prc = TRUE)

plot(roc)

plot(prc, prc = TRUE)
}
