% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROCcurve.R
\name{ROCcurve}
\alias{ROCcurve}
\title{ROC and PRC curves}
\usage{
ROCcurve(
  y,
  X = NULL,
  model = NULL,
  plot = FALSE,
  optPoint = TRUE,
  grid = FALSE,
  grid_lty = 3,
  grid_lwd = 1.5,
  grid_col = "lightgray",
  midline = TRUE,
  midline_lty = 2,
  midline_lwd = 2,
  midline_col = "red",
  pt_pch = 23,
  pt_border = "black",
  pt_col = "green",
  thresh = TRUE,
  roc_lty = 1,
  roc_lwd = 2,
  roc_col = "black",
  prc = FALSE
)
}
\arguments{
\item{y}{Vector with data for binary outcome}

\item{X}{Data matrix or vector of predictors}

\item{model}{Optional - provide logistic regression model}

\item{plot}{Logical}

\item{optPoint}{Logical}

\item{grid}{Logical}

\item{grid_lty}{Numeric}

\item{grid_lwd}{Numeric}

\item{grid_col}{Character}

\item{midline}{Logical}

\item{midline_lty}{Numeric}

\item{midline_lwd}{Numeric}

\item{midline_col}{Character}

\item{pt_pch}{Numeric}

\item{pt_border}{Character}

\item{pt_col}{Character}

\item{thresh}{Logical}

\item{roc_lty}{Numeric}

\item{roc_lwd}{Numeric}

\item{roc_col}{Character}

\item{prc}{Logical}
}
\value{
An ROCcurve object or plot
}
\description{
These are ROC and PRC curves
}
\examples{
1 + 1
}
