% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROCcurve.R
\name{summary.ROCcurve}
\alias{summary.ROCcurve}
\title{Summarize optimal ROC and PRC results}
\usage{
\method{summary}{ROCcurve}(object, type = c("roc", "prc"), ...)
}
\arguments{
\item{object}{ROCcurve object}

\item{type}{Indicate \code{"roc"}, \code{"prc"}, or both}

\item{...}{Additional arguments}
}
\value{
The optimal parameter values for the ROC or PRC curve
}
\description{
Summarize optimal ROC and PRC results
}
\examples{
1 + 1
}
