% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresh.R
\name{thresh}
\alias{thresh}
\title{Find threshold scale value based on ROC or PRC}
\usage{
thresh(x, y, prc = FALSE, scale.max = "default", cutoff = ">", round = 5)
}
\arguments{
\item{x}{vector of predictor values}

\item{y}{vector of binary outcome values}

\item{prc}{Logical, determines whether to base threshold on PRC}

\item{scale.max}{largest value of the predictor, or leave as default}

\item{cutoff}{Character string. Define how to classify cases relative to the
predicted probability cutoff}

\item{round}{how many decimal places to round the cutoff and predicted
probabilities}
}
\value{
A value representing the largest value that would not be classified
in the positive class
}
\description{
Find threshold scale value based on ROC or PRC
}
\examples{
1 + 1
}
